moved kompasText class to text_styles
